amundsen/frontend/requirements.txt
# Copyright Contributors to the Amundsen project.
# SPDX-License-Identifier: Apache-2.0

responses==0.12.1
SQLAlchemy==1.3.23
jira==3.0.1
asana==0.10.3
retrying>=1.3.3,<2.0

# Backport of PEP 557, Data Classes for Python 3.6.
# License: Apache 2.0
# Upstream url: https://pypi.org/project/dataclasses/
dataclasses==0.8; python_version < '3.7'

# extra files for gauth
flask-oidc==1.4.0
flaskoidc==0.2.3
authlib==0.15.4

